Finally got around to pushing my latest project to github. This is a arduino sketch to control my model rocket.
Current Functionality is it starts in a initial safety state so the parachute black powder charge doesn't go off inadvertantly,
you press a button and it gets ready to detect flight using the accelerometer readings,
once in flight it gets data of the MPU and altimeter.
Then the altimeter detects the apogee of the flight path.
Once apogee is detected, high voltage is applied to the parachute charge arduino pin.

This is totally not done yet. I need to take some time and write code to log the sensor data to a file stored on a SD card for processing.